% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{addSeasonalRegression}
\alias{addSeasonalRegression}
\title{Add a cosine model regression to plot.}
\usage{
addSeasonalRegression(.plot)
}
\arguments{
\item{.plot}{A ggplot plot from plotConsumption}
}
\value{
A ggplot object with the linear regression added
}
\description{
This is aimed to show wether there is seasonality on the energy consumption.
}
