% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{addSeasonBars}
\alias{addSeasonBars}
\title{Add season boxes to the bottom to a plot}
\usage{
addSeasonBars(.plot)
}
\arguments{
\item{.plot}{A ggplot2 object from consumptionPlot}
}
\value{
A ggplot object with season boxes
}
\description{
Add season boxes to the bottom to a plot
}
