% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{anonymise}
\alias{anonymise}
\title{Anonymise data}
\usage{
anonymise(.data)
}
\arguments{
\item{.data}{A tibble}
}
\value{
A tibble without sensible information
}
\description{
Substitutes Account Id information with random names
also adds some jitter noise to the data.
Useful if you are planning to publish a plot
somwehre on the internet
}
