% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{connect}
\alias{connect}
\title{Connects to My Ovo Rest service.}
\usage{
connect(username, password)
}
\arguments{
\item{username}{OvO username}

\item{password}{OvO password}
}
\value{
A file with the connection cookie.
}
\description{
Connects to My Ovo Rest service.
}
