% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect from the OvO account}
\usage{
disconnect(cookie.path)
}
\arguments{
\item{cookie.path}{A file with the cookie previously stablished}
}
\description{
Removes the cookie file
TODO: Use the API to disconnect
}
