% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{filterByUtility}
\alias{filterByUtility}
\title{Extracts the rows pertaining to the given utility}
\usage{
filterByUtility(.data, .utility)
}
\arguments{
\item{.data}{The data tibble}

\item{.utility}{the utility name we are interested in}
}
\value{
A new data tibble
}
\description{
Extracts the rows pertaining to the given utility
}
