% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{filterByYears}
\alias{filterByYears}
\title{Extract the rows with only the given years}
\usage{
filterByYears(.data, .years)
}
\arguments{
\item{.data}{The data tibble}

\item{.years}{A vector with years}
}
\value{
A new data tibble
}
\description{
Extract the rows with only the given years
}
