% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{getAccountIds}
\alias{getAccountIds}
\title{Query the OvO API for the account IDs}
\usage{
getAccountIds(cookie.path)
}
\arguments{
\item{cookie.path}{A file with the cookie previously stablished}
}
\value{
A vector of account IDs
}
\description{
Normally, it's only one.
}
