% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{getData}
\alias{getData}
\title{Retrives the account data from OvO and composes a data tibble}
\usage{
getData(cookie.path)
}
\arguments{
\item{cookie.path}{A file with the cookie previously stablished}
}
\value{
A data tibble with all the information
}
\description{
This functions uses the cookie file to connect to the REST
API. Navigates trough the complex json structure and composes
a data tbble with all the consumption and cost information available.
}
