% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{getMonthlyData}
\alias{getMonthlyData}
\title{Get the data for a given month and account}
\usage{
getMonthlyData(accId, date, cookie.path)
}
\arguments{
\item{cookie.path}{A file with the cookie previously stablished}

\item{accID}{Account ID}

\item{data}{A date object with the monnth}
}
\value{
A data tibble for that month and account id
}
\description{
The returned tibble will have this shape
Consumption Date       Cost  RateAnytime RateStanding Utility
\if{html}{\out{<dbl>}} \if{html}{\out{<date>}}     \if{html}{\out{<chr>}}       \if{html}{\out{<dbl>}}        \if{html}{\out{<dbl>}} \if{html}{\out{<fct>}}
1        9.87 2023-04-01 3.15        0.319        0.472 electricity
2        5.28 2023-04-02 1.69        0.319        0.472 electricity
...
}
