% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{getPlans}
\alias{getPlans}
\title{Return the account plans in a list}
\usage{
getPlans(accId, cookie.path)
}
\arguments{
\item{accId}{Account ID}

\item{cookie.path}{A file with the cookie previously stablished}
}
\value{
A list with the plans
}
\description{
Return the account plans in a list
}
