% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{julDates}
\alias{julDates}
\title{Create a table that maps current year's julian dates
to human readable month names.
This is used for relabeling the x axis ticks}
\usage{
julDates()
}
\description{
Create a table that maps current year's julian dates
to human readable month names.
This is used for relabeling the x axis ticks
}
