% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{makeConsumptionPlot}
\alias{makeConsumptionPlot}
\title{Make a Consumption Plot}
\usage{
makeConsumptionPlot(.d, .u, .y, .h, .rmw, .r, .s)
}
\arguments{
\item{.d}{tibble data frome from getData}

\item{.u}{string with the utility name}

\item{.y}{vector of years}

\item{.h}{boolean for hidding information or not}

\item{.rmw}{numeric with a rolling mean window in days}

\item{.r}{boolean for adding a regression line or not}

\item{.s}{boolean for adding a season bar or not}
}
\value{
a ggplot2 object
}
\description{
Make a Consumption Plot
}
