% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{makeCostPlot}
\alias{makeCostPlot}
\title{Make a Cost Plot}
\usage{
makeCostPlot(.d, .u, .y, .h, .rmw)
}
\arguments{
\item{.d}{tibble data frome from getData}

\item{.u}{string with the utility name}

\item{.y}{vector of years}

\item{.h}{boolean for hidding information or not}

\item{.rmw}{numeric with a rolling mean window in days}
}
\value{
a ggplot2 object
}
\description{
Make a Cost Plot
}
