% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{makeFunConditional}
\alias{makeFunConditional}
\title{Transforms a function into a conditional function}
\usage{
makeFunConditional(.f)
}
\arguments{
\item{.f}{A function with one parameter .f(.d)}
}
\value{
A new function with two paramters .f(.d, .b)
}
\description{
Receives a function with one paramter, and transforms it
into a function with two paramters (adds one paramter).
The new paramter is a boolean. If its value is TRUE
executes the function, otherwise returns the original
data.
}
