% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotConsumption}
\alias{plotConsumption}
\title{Plot the Consumption}
\usage{
plotConsumption(.data)
}
\arguments{
\item{.data}{A data tibble previously treated}
}
\value{
A ggplot object with the consumption
}
\description{
You need to filter by utility,
select the years and perhaps
calculate a rolling mean before using this function.
}
