% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCost}
\alias{plotCost}
\title{Plot the Costs}
\usage{
plotCost(.data)
}
\arguments{
\item{.data}{A data tibble previously treated}
}
\value{
A ggplot object with the cost
}
\description{
You need to filter by utility,
select the years and perhaps
calculate a rolling mean before using this function.
}
\details{
The actual consumption is plotted in red
and the standing charge in blue.
}
