% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{seasonDates}
\alias{seasonDates}
\title{Get the start and end of the 4 seasons.}
\usage{
seasonDates(.data)
}
\arguments{
\item{.data}{A data tibble}
}
\value{
A data tibble with the season information
}
\description{
Also calculate a ymax and ymin limits.
With this information one can draw a box at the
bottom of a plot representing the seasons.
}
